refactor(docs/animationhelpers): Create hold-tap subfolders
