+++ docs/src/animationhelpers/src/hold_tap/balanced/balanced_hold_tap_up.ts
@@ -1,6 +1,6 @@
 import { makeProject } from "@motion-canvas/core/lib";
 
-import balanced_hold_tap_up from "../scenes/hold_tap/balanced_hold_tap_up?scene";
+import balanced_hold_tap_up from "../../scenes/hold_tap/balanced/balanced_hold_tap_up?scene";
 
 export default makeProject({
   scenes: [balanced_hold_tap_up],
