+++ docs/src/animationhelpers/src/hold_tap/balanced/balanced_other_key_up.ts
@@ -1,6 +1,6 @@
 import { makeProject } from "@motion-canvas/core/lib";
 
-import balanced_other_key_up from "../scenes/hold_tap/balanced_other_key_up?scene";
+import balanced_other_key_up from "../../scenes/hold_tap/balanced/balanced_other_key_up?scene";
 
 export default makeProject({
   scenes: [balanced_other_key_up],
