+++ docs/src/animationhelpers/src/hold_tap/tap_preferred/tap_preferred_hold_tap_up.ts
@@ -1,6 +1,6 @@
 import { makeProject } from "@motion-canvas/core/lib";
 
-import tap_preferred_hold_tap_up from "../scenes/hold_tap/tap_preferred_hold_tap_up?scene";
+import tap_preferred_hold_tap_up from "../../scenes/hold_tap/tap_preferred/tap_preferred_hold_tap_up?scene";
 
 export default makeProject({
   scenes: [tap_preferred_hold_tap_up],
