+++ docs/src/animationhelpers/src/hold_tap/tap_unless_interrupted/tap_unless_interrupted.ts
@@ -1,6 +1,6 @@
 import { makeProject } from "@motion-canvas/core/lib";
 
-import tap_unless_interrupted from "../scenes/hold_tap/tap_unless_interrupted?scene";
+import tap_unless_interrupted from "../../scenes/hold_tap/tap_unless_interrupted/tap_unless_interrupted?scene";
 
 export default makeProject({
   scenes: [tap_unless_interrupted],
