+++ docs/src/animationhelpers/src/scenes/hold_tap/balanced/balanced_hold_tap_up.meta
