+++ docs/src/animationhelpers/src/scenes/hold_tap/tap_unless_interrupted/tap_unless_interrupted.tsx
@@ -2,8 +2,8 @@ import { makeScene2D } from "@motion-canvas/2d/lib/scenes";
 import { Text } from "@motion-canvas/2d/lib/components";
 import { createRef, makeRefs } from "@motion-canvas/core/lib/utils";
 import { all, chain, delay, waitFor } from "@motion-canvas/core/lib/flow";
-import Key, { KeyTravel } from "../../components/Key";
-import Output from "../../components/Output";
+import Key, { KeyTravel } from "../../../components/Key";
+import Output from "../../../components/Output";
 import { linear } from "@motion-canvas/core/lib/tweening";
 
 export default makeScene2D(function* (view) {
