+++ docs/src/animationhelpers/vite.config.ts
@@ -7,11 +7,11 @@ export default defineConfig({
       project: [
         "./src/hold_tap/hold_tap_comparison.ts",
         "./src/hold_tap/hold_tap_interrupted.ts",
-        "./src/hold_tap/balanced_other_key_up.ts",
-        "./src/hold_tap/balanced_hold_tap_up.ts",
-        "./src/hold_tap/tap_preferred_hold_tap_up.ts",
-        "./src/hold_tap/tap_unless_interrupted.ts",
-        "./src/hold_tap/tap_unless_interrupted_invoke_hold.ts",
+        "./src/hold_tap/balanced/balanced_other_key_up.ts",
+        "./src/hold_tap/balanced/balanced_hold_tap_up.ts",
+        "./src/hold_tap/tap_preferred/tap_preferred_hold_tap_up.ts",
+        "./src/hold_tap/tap_unless_interrupted/tap_unless_interrupted.ts",
+        "./src/hold_tap/tap_unless_interrupted/tap_unless_interrupted_invoke_hold.ts",
       ],
     }),
   ],
